Move object and collection structs to protocol folder
