+++ crates/apub/src/collections/mod.rs
@@ -1,10 +1,9 @@
-use crate::objects::community::ApubCommunity;
 use lemmy_websocket::LemmyContext;
 
-pub(crate) mod community_followers;
+use crate::objects::community::ApubCommunity;
+
 pub(crate) mod community_moderators;
 pub(crate) mod community_outbox;
-pub(crate) mod user_outbox;
 
 /// Put community in the data, so we dont have to read it again from the database.
 pub(crate) struct CommunityContext(pub ApubCommunity, pub LemmyContext);
