+++ crates/apub/src/protocol/collections/group_moderators.rs
@@ -1,0 +1,12 @@
+use crate::{fetcher::object_id::ObjectId, objects::person::ApubPerson};
+use activitystreams::collection::kind::OrderedCollectionType;
+use serde::{Deserialize, Serialize};
+use url::Url;
+
+#[derive(Clone, Debug, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct GroupModerators {
+  pub(crate) r#type: OrderedCollectionType,
+  pub(crate) id: Url,
+  pub(crate) ordered_items: Vec<ObjectId<ApubPerson>>,
+}
