+++ crates/apub/src/protocol/collections/group_outbox.rs
@@ -1,0 +1,13 @@
+use crate::activities::post::create_or_update::CreateOrUpdatePost;
+use activitystreams::collection::kind::OrderedCollectionType;
+use serde::{Deserialize, Serialize};
+use url::Url;
+
+#[derive(Clone, Debug, Deserialize, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct GroupOutbox {
+  pub(crate) r#type: OrderedCollectionType,
+  pub(crate) id: Url,
+  pub(crate) total_items: i32,
+  pub(crate) ordered_items: Vec<CreateOrUpdatePost>,
+}
