+++ crates/apub/src/protocol/collections/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod group_followers;
+pub(crate) mod group_moderators;
+pub(crate) mod group_outbox;
+pub(crate) mod person_outbox;
