+++ crates/apub/src/protocol/objects/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod chat_message;
+pub(crate) mod group;
+pub(crate) mod note;
+pub(crate) mod page;
+pub(crate) mod tombstone;
